+++ examples/tempura/src/views/pages/homeData.js
@@ -1,0 +1,8 @@
+module.exports = {
+  title: 'Tempura',
+  persons: [
+    { name: 'Nils', age: 20 },
+    { name: 'Robert', age: 10 },
+    { name: 'Margaret', age: 40 },
+  ],
+};